Add some stopwatches to ovn-nbctl.
For one-off function calls, we check the maximum time recorded by the
stopwatch since it is the only time recorded.

For functions called repeatedly, we print the max and average on each
iteration. This way, we can see the change over time in the maximum. And
if the maximum stays the same but the average increases, we know that
something between the current maximum and average was recorded.